+++ ReactAndroid/src/main/java/com/facebook/yoga/YogaConfig.java
@@ -106,7 +106,7 @@ public class YogaConfig {
   }
 
   @DoNotStrip
-  private final YogaNode cloneNode(YogaNode oldNode, YogaNode parent, int childIndex) {
-    return mYogaNodeCloneFunction.cloneNode(oldNode, parent, childIndex);
+  private final YogaNodeJNI cloneNode(YogaNodeJNI oldNode, YogaNodeJNI parent, int childIndex) {
+    return (YogaNodeJNI) mYogaNodeCloneFunction.cloneNode(oldNode, parent, childIndex);
   }
 }
